all: add a code of conduct
Copy the code of conduct from mellium.im/xmpp. This provides guidance
for project maintainers and contributors alike and gives us an outline
for how to handle problems and conflict. While it won't necessarily
prevent or mitigate conflicts itself, it can ensure that we treat
everyone by the same set of guidelines and help avoid favoritism or
unconcious biases.