docs: consolidate docs in one place
This patch consolidates all documenation and GitHub specific config
files such as FUNDING.yml into the docs/ tree.
It also removes docs that are duplicates of the same or similar docs in
the main repo and replaces them with links to the canonical docs so that
changes don't have to be made across repos in the future.

Signed-off-by: Sam Whited <sam@samwhited.com>