sasl: error if CB is used without tls-unique data
An issue was discovered where the Prosody XMPP server would advertise
support for channel binding mechanisms in the SCRAM family even if TLS
1.3 was used (for which the tls-unique channel binding type is not
defined). This library would fail if no TLS connection state was
provided, but would hash in invalid channel binding data if a connection
state was provided but no TLS unique data existed.  If both the client
and server had a similar bug, it could possibly lead to a security issue
where authentication completes but the channel binding is not valid.

For more information see https://issues.prosody.im/1542 and
https://mellium.im/issue/46 (mellium/xmpp#46).

Signed-off-by: Sam Whited <sam@samwhited.com>